Move class to break class cycle
